requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.3
Flask-PyMongo==3.0.1
Flask-SocketIO==5.6.1
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
http_ece==1.2.1
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mongomock==4.3.0
multidict==6.7.1
packaging==26.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
py-vapid==1.9.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pyparsing==3.3.2
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2025.2
pywebpush==2.3.0
requests==2.32.5
sentinels==1.1.1
simple-websocket==1.1.0
soupsieve==2.8.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.6
wsproto==1.3.2
yarl==1.22.0
gunicorn
